refactor: Extend shared graph workspace for undirected/weighted reuse
- Move weightedAdjacency/edgeList into graph.js as core helpers
  (shortestPath.js re-exports them; its imports are unchanged).
- floating-edge: draw the mid-edge travel arrow only when a direction
  is set (travelTo), so undirected uses get no arrow; add a 'reject'
  edge color.
- graph-menu: add a hideDirected prop to omit the Directed toggle.
